rebase on app-32 and consolidate ahsh and action button ssection into one
